Major update and rewrite to everything.
Rewrote entire source code of bot to be in line with accepted practices vis a vis the particular wrapper chosen for this project. 

Reformatted files to meet PEP standards across-the-board, and properly secured key in token file instead of unknown dotenv. 

Functions rewritten to be pythonic in nature.

Properly used lists to store XML data as opposed to storing in local file during runtime, in line with programming best practices.

Configured debug-level logging as opposed to the nonexistent logging that prior existed to ensure that errors can be properly caught and tracked.

Changed web interactions wrapper to `aiohttp` to properly enable asynchronous web requests, along with usage of async context managers and loops to make sure that requests were handled in a non-blocking manner.